fix(material-experimental/mdc-slide-toggle): initial checked a… (#17991)
The initial `checked` and `disabled` state of the slide-toggle
is not reflected visually. This is because we sync the disabled
and checked state after the foundation has been created in the
`ngAfterViewInit` lifecycle hook.

The foundation then tries to set the classes on the MDC switch
element using the adapter. The adapter just updates an object
literal that will be passed through `[ngClass]` and relies on
Angular to refresh the view.

This won't work in Ivy because marking a component as dirty
from within the `ngAfterViewInit` lifecycle hook is a noop. See: FW-1354.

Just using the native element (like we do in other MDC components)
solves this problem and also avoids unnecessary change detections.