fix(material-experimental/mdc-chips): removed chip still focus… (#18083)
When an MDC chip is removed it gets transitioned to `width: 0`, but it's up to the consumer to remove it from the DOM. This means that it's possible for the user to tab into it. These changes add `display: none` after the animation is done so the element becomes non-focusable.